// generated/typescript/models/Clusters.ts
/**
 * Severalnines ClusterControl
 * This is a ...
 *
 * OpenAPI spec version: 1.0.0
 * Contact: sales@severalnines.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ClustersAccount } from './ClustersAccount';
import { ClustersConfiguration } from './ClustersConfiguration';
import { ClustersDatabase } from './ClustersDatabase';
import { HttpFile } from '../http/http';

export class Clusters {
    'operation': ClustersOperationEnum;
    'clusterId'?: number;
    'clusterName'?: string;
    'withHosts'?: boolean;
    'withSheetInfo'?: boolean;
    'configuration'?: Array<ClustersConfiguration>;
    'account'?: ClustersAccount;
    'filterStrings'?: string;
    'limit'?: number;
    'offset'?: number;
    'database'?: ClustersDatabase;
    'nodes'?: Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "operation",
            "baseName": "operation",
            "type": "ClustersOperationEnum",
            "format": ""
        },
        {
            "name": "clusterId",
            "baseName": "cluster_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "clusterName",
            "baseName": "cluster_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "withHosts",
            "baseName": "with_hosts",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "withSheetInfo",
            "baseName": "with_sheet_info",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "configuration",
            "baseName": "configuration",
            "type": "Array<ClustersConfiguration>",
            "format": ""
        },
        {
            "name": "account",
            "baseName": "account",
            "type": "ClustersAccount",
            "format": ""
        },
        {
            "name": "filterStrings",
            "baseName": "filter_strings",
            "type": "string",
            "format": ""
        },
        {
            "name": "limit",
            "baseName": "limit",
            "type": "number",
            "format": ""
        },
        {
            "name": "offset",
            "baseName": "offset",
            "type": "number",
            "format": ""
        },
        {
            "name": "database",
            "baseName": "database",
            "type": "ClustersDatabase",
            "format": ""
        },
        {
            "name": "nodes",
            "baseName": "nodes",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Clusters.attributeTypeMap;
    }

    public constructor() {
    }
}


export type ClustersOperationEnum = "getallclusterinfo" | "getclusterinfo" | "getConfig" | "setConfig" | "createAccount" | "getAccounts" | "deleteAccount" | "grantPrivileges" | "revokePrivileges" | "revokeAllPrivileges" | "getSqlProcesses" | "getTopQueries" | "createDatabase" | "ping" | "availableUpgrades" ;