// generated/typescript/models/Config.ts
/**
 * Severalnines ClusterControl
 * This is a ...
 *
 * OpenAPI spec version: 1.0.0
 * Contact: sales@severalnines.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ConfigConfiguration } from './ConfigConfiguration';
import { ConfigLdapConfiguration } from './ConfigLdapConfiguration';
import { HttpFile } from '../http/http';

export class Config {
    'operation': ConfigOperationEnum;
    'hostname'?: string;
    'port'?: number;
    'configuration'?: Array<ConfigConfiguration>;
    'ldapConfiguration'?: ConfigLdapConfiguration;
    'licensedata'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "operation",
            "baseName": "operation",
            "type": "ConfigOperationEnum",
            "format": ""
        },
        {
            "name": "hostname",
            "baseName": "hostname",
            "type": "string",
            "format": ""
        },
        {
            "name": "port",
            "baseName": "port",
            "type": "number",
            "format": ""
        },
        {
            "name": "configuration",
            "baseName": "configuration",
            "type": "Array<ConfigConfiguration>",
            "format": ""
        },
        {
            "name": "ldapConfiguration",
            "baseName": "ldap_configuration",
            "type": "ConfigLdapConfiguration",
            "format": ""
        },
        {
            "name": "licensedata",
            "baseName": "licensedata",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Config.attributeTypeMap;
    }

    public constructor() {
    }
}


export type ConfigOperationEnum = "getConfig" | "setConfig" | "unsetConfig" | "getLdapConfig" | "setLdapConfig" | "setLicense" | "getLicense" | "listTemplates" ;